Target framework now 4.8 Added enum for script names Changed read / writes to use a generic method Using nint instead of IntPtr throughout the codebase

+++ SekiroTool/Interfaces/ITargetService.cs
@@ -4,7 +4,7 @@ public interface ITargetService
 {
     void ToggleTargetHook(bool isEnabled);
 
-    ulong GetTargetChrIns();
+    nint GetTargetChrIns();
     
     void SetHp(int hp);
     int GetCurrentHp();
